Be sure to check the exit code of certbot
If we use set -e but no -o pipefail with a pipe, the last command exit code is
used by set -e.

certbot | grep -v something

If the grep exit with a non-zero exit code, set -e stop the execution of the
script.  We don't care about that grep, so we now use the PIPESTATUS.