certbot: Properly evaluate when apache is installed
Checking the existence of /etc/apache2 is not enough as a condition to
validate the presence of apache.

Indeed, some packages (including certbot!!!), put some files in
/etc/apache2/conf-available even if apache isn't installed.

In those cases, the check is not correct, and we'll enter in the apache
block, and fail when we try to enable the configuration.

With this commit, we now validate the presence apache with the presence
of /usr/sbin/apachectl