refactor(redis): Avoid using set_fact when possible
Variables that are set as facts are cached at the host level, which can
result in problems if the role is called multiple time in a playbook.

Some of the variables were only used in a single task so they were moved
as task level variables (some templated).

Others apply to the whole role and where moved to variable files.
(Either "vars/main.yml" that is statically read when the role is read by
ansible, or "vars/<file>" that is dynamically read by an
`ansible.builtin.include_vars` when they are conditionnal overrides.)

(Note: In all cases variables set by `set_vars` already overrode
inventory and play vars, which is still the case after this change.)